requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
feedparser==6.0.11
google-genai>=1.0.0
httpx~=0.25.2
python-dotenv==1.0.1
pytz==2024.1
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
